+++ BlanktopiaItems/src/main/kotlin/me/weiwen/blanktopia/actions/ExperienceBoostAction.kt
@@ -3,7 +3,7 @@ package me.weiwen.blanktopia.actions
 import me.weiwen.blanktopia.addExperienceBoost
 import org.bukkit.entity.Player
 
-class ExperienceBoostAction(private val multiplier: Double, private val ticks: Long) : Action {
+class ExperienceBoostAction(private val multiplier: Double, private val ticks: Int) : Action {
     override fun run(player: Player) {
         player.addExperienceBoost(multiplier, ticks)
     }
